requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
asyncio>=3.4.3
fastapi>=0.68.0
uvicorn>=0.15.0 